+++ apps/dashboard/src/app/team/[team_slug]/[project_slug]/engine/(instance)/[engineId]/access-tokens/page.tsx
@@ -1,8 +1,8 @@
 "use client";
 
 import { WithEngineInstance } from "components/engine/EnginePageLayout";
-import { EngineAccessTokens } from "components/engine/permissions/engine-access-tokens";
 import type { EngineInstancePageProps } from "../types";
+import { EngineAccessTokens } from "./components/engine-access-tokens";
 
 export default function Page(props: EngineInstancePageProps) {
   const { params } = props;
