+++ apps/dashboard/src/app/team/[team_slug]/[project_slug]/engine/(instance)/[engineId]/alerts/components/RecentEngineAlerts.stories.tsx
@@ -3,8 +3,8 @@ import { Toaster } from "sonner";
 import {
   createEngineAlertRuleStub,
   createEngineAlertStub,
-} from "../../../stories/stubs";
-import { BadgeContainer, mobileViewport } from "../../../stories/utils";
+} from "stories/stubs";
+import { BadgeContainer, mobileViewport } from "stories/utils";
 import { RecentEngineAlertsSectionUI } from "./RecentEngineAlerts";
 
 const meta = {
