+++ apps/dashboard/src/app/team/[team_slug]/[project_slug]/engine/(instance)/[engineId]/explorer/components/swagger-ui.css
@@ -525,27 +525,27 @@
 
 @-webkit-keyframes rotation {
   to {
-    -webkit-transform: rotate(1turn) !important;
-    transform: rotate(1turn) !important;
+    -webkit-transform: rotate(1turn);
+    transform: rotate(1turn);
   }
 }
 
 @keyframes rotation {
   to {
-    -webkit-transform: rotate(1turn) !important;
-    transform: rotate(1turn) !important;
+    -webkit-transform: rotate(1turn);
+    transform: rotate(1turn);
   }
 }
 
 @-webkit-keyframes blinker {
   50% {
-    opacity: 0 !important;
+    opacity: 0;
   }
 }
 
 @keyframes blinker {
   50% {
-    opacity: 0 !important;
+    opacity: 0;
   }
 }
 
@@ -603,33 +603,33 @@
 
 @-webkit-keyframes pulse {
   0% {
-    color: #fff !important;
-    background: #4990e2 !important;
-    box-shadow: 0 0 0 0 rgba(73, 144, 226, 0.8) !important;
+    color: #fff;
+    background: #4990e2;
+    box-shadow: 0 0 0 0 rgba(73, 144, 226, 0.8);
   }
   70% {
-    box-shadow: 0 0 0 5px rgba(73, 144, 226, 0) !important;
+    box-shadow: 0 0 0 5px rgba(73, 144, 226, 0);
   }
   to {
-    color: #fff !important;
-    background: #4990e2 !important;
-    box-shadow: 0 0 0 0 rgba(73, 144, 226, 0) !important;
+    color: #fff;
+    background: #4990e2;
+    box-shadow: 0 0 0 0 rgba(73, 144, 226, 0);
   }
 }
 
 @keyframes pulse {
   0% {
-    color: #fff !important;
-    background: #4990e2 !important;
-    box-shadow: 0 0 0 0 rgba(73, 144, 226, 0.8) !important;
+    color: #fff;
+    background: #4990e2;
+    box-shadow: 0 0 0 0 rgba(73, 144, 226, 0.8);
   }
   70% {
-    box-shadow: 0 0 0 5px rgba(73, 144, 226, 0) !important;
+    box-shadow: 0 0 0 5px rgba(73, 144, 226, 0);
   }
   to {
-    color: #fff !important;
-    background: #4990e2 !important;
-    box-shadow: 0 0 0 0 rgba(73, 144, 226, 0) !important;
+    color: #fff;
+    background: #4990e2;
+    box-shadow: 0 0 0 0 rgba(73, 144, 226, 0);
   }
 }
 
@@ -766,48 +766,48 @@
 @-webkit-keyframes shake {
   10%,
   90% {
-    -webkit-transform: translate3d(-1px, 0, 0) !important;
-    transform: translate3d(-1px, 0, 0) !important;
+    -webkit-transform: translate3d(-1px, 0, 0);
+    transform: translate3d(-1px, 0, 0);
   }
   20%,
   80% {
-    -webkit-transform: translate3d(2px, 0, 0) !important;
-    transform: translate3d(2px, 0, 0) !important;
+    -webkit-transform: translate3d(2px, 0, 0);
+    transform: translate3d(2px, 0, 0);
   }
   30%,
   50%,
   70% {
-    -webkit-transform: translate3d(-4px, 0, 0) !important;
-    transform: translate3d(-4px, 0, 0) !important;
+    -webkit-transform: translate3d(-4px, 0, 0);
+    transform: translate3d(-4px, 0, 0);
   }
   40%,
   60% {
-    -webkit-transform: translate3d(4px, 0, 0) !important;
-    transform: translate3d(4px, 0, 0) !important;
+    -webkit-transform: translate3d(4px, 0, 0);
+    transform: translate3d(4px, 0, 0);
   }
 }
 
 @keyframes shake {
   10%,
   90% {
-    -webkit-transform: translate3d(-1px, 0, 0) !important;
-    transform: translate3d(-1px, 0, 0) !important;
+    -webkit-transform: translate3d(-1px, 0, 0);
+    transform: translate3d(-1px, 0, 0);
   }
   20%,
   80% {
-    -webkit-transform: translate3d(2px, 0, 0) !important;
-    transform: translate3d(2px, 0, 0) !important;
+    -webkit-transform: translate3d(2px, 0, 0);
+    transform: translate3d(2px, 0, 0);
   }
   30%,
   50%,
   70% {
-    -webkit-transform: translate3d(-4px, 0, 0) !important;
-    transform: translate3d(-4px, 0, 0) !important;
+    -webkit-transform: translate3d(-4px, 0, 0);
+    transform: translate3d(-4px, 0, 0);
   }
   40%,
   60% {
-    -webkit-transform: translate3d(4px, 0, 0) !important;
-    transform: translate3d(4px, 0, 0) !important;
+    -webkit-transform: translate3d(4px, 0, 0);
+    transform: translate3d(4px, 0, 0);
   }
 }
 
@@ -1521,27 +1521,27 @@
 
 @-webkit-keyframes scaleUp {
   0% {
-    -webkit-transform: scale(0.8) !important;
-    transform: scale(0.8) !important;
-    opacity: 0 !important;
+    -webkit-transform: scale(0.8);
+    transform: scale(0.8);
+    opacity: 0;
   }
   to {
-    -webkit-transform: scale(1) !important;
-    transform: scale(1) !important;
-    opacity: 1 !important;
+    -webkit-transform: scale(1);
+    transform: scale(1);
+    opacity: 1;
   }
 }
 
 @keyframes scaleUp {
   0% {
-    -webkit-transform: scale(0.8) !important;
-    transform: scale(0.8) !important;
-    opacity: 0 !important;
+    -webkit-transform: scale(0.8);
+    transform: scale(0.8);
+    opacity: 0;
   }
   to {
-    -webkit-transform: scale(1) !important;
-    transform: scale(1) !important;
-    opacity: 1 !important;
+    -webkit-transform: scale(1);
+    transform: scale(1);
+    opacity: 1;
   }
 }
