+++ apps/dashboard/src/app/team/[team_slug]/[project_slug]/engine/(instance)/[engineId]/relayers/page.tsx
@@ -1,8 +1,8 @@
 "use client";
 
 import { WithEngineInstance } from "components/engine/EnginePageLayout";
-import { EngineRelayer } from "components/engine/relayer/engine-relayer";
 import type { EngineInstancePageProps } from "../types";
+import { EngineRelayer } from "./components/engine-relayer";
 
 export default function Page(props: EngineInstancePageProps) {
   const { params } = props;
