+++ apps/dashboard/src/pages/dashboard/engine/[engineId]/access-tokens.tsx
@@ -1,5 +1,5 @@
+import { EngineAccessTokens } from "app/team/[team_slug]/[project_slug]/engine/(instance)/[engineId]/access-tokens/components/engine-access-tokens";
 import { createEnginePage } from "components/engine/EnginePage";
-import { EngineAccessTokens } from "components/engine/permissions/engine-access-tokens";
 
 export default createEnginePage(({ instance }) => (
   <EngineAccessTokens instanceUrl={instance.url} />
