+++ apps/dashboard/src/pages/dashboard/engine/[engineId]/admins.tsx
@@ -1,5 +1,5 @@
+import { EngineAdmins } from "app/team/[team_slug]/[project_slug]/engine/(instance)/[engineId]/admins/components/engine-admins";
 import { createEnginePage } from "components/engine/EnginePage";
-import { EngineAdmins } from "components/engine/permissions/engine-admins";
 
 export default createEnginePage(({ instance }) => (
   <EngineAdmins instanceUrl={instance.url} />
