+++ apps/dashboard/src/pages/dashboard/engine/[engineId]/alerts.tsx
@@ -1,5 +1,5 @@
+import { EngineAlertsPage } from "app/team/[team_slug]/[project_slug]/engine/(instance)/[engineId]/alerts/components/EngineAlertsPage";
 import { createEnginePage } from "components/engine/EnginePage";
-import { EngineAlertsPage } from "../../../../components/engine/alerts/EngineAlertsPage";
 
 export default createEnginePage(({ instance }) => (
   <EngineAlertsPage instance={instance} />
