+++ apps/dashboard/src/pages/dashboard/engine/[engineId]/configuration.tsx
@@ -1,5 +1,5 @@
+import { EngineConfiguration } from "app/team/[team_slug]/[project_slug]/engine/(instance)/[engineId]/configuration/components/engine-configuration";
 import { createEnginePage } from "components/engine/EnginePage";
-import { EngineConfiguration } from "components/engine/configuration/engine-configuration";
 
 export default createEnginePage(({ instance }) => (
   <EngineConfiguration instance={instance} />
