+++ apps/dashboard/src/pages/dashboard/engine/[engineId]/metrics.tsx
@@ -1,5 +1,5 @@
+import { EngineSystemMetrics } from "app/team/[team_slug]/[project_slug]/engine/(instance)/[engineId]/metrics/components/EngineSystemMetrics";
 import { createEnginePage } from "components/engine/EnginePage";
-import { EngineSystemMetrics } from "components/engine/system-metrics";
 
 export default createEnginePage(({ instance }) => (
   <EngineSystemMetrics instance={instance} />
