+++ apps/dashboard/src/pages/dashboard/engine/[engineId]/webhooks.tsx
@@ -1,5 +1,5 @@
+import { EngineWebhooks } from "app/team/[team_slug]/[project_slug]/engine/(instance)/[engineId]/webhooks/components/engine-webhooks";
 import { createEnginePage } from "components/engine/EnginePage";
-import { EngineWebhooks } from "components/engine/configuration/engine-webhooks";
 
 export default createEnginePage(({ instance }) => (
   <EngineWebhooks instanceUrl={instance.url} />
